Cleanup in code

+++ PropagationPuzzle/Assets/Scripts/Sound/SoundEmitter.cs
@@ -84,9 +84,8 @@ public class SoundEmitter : MonoBehaviour
         {
 
             //Calculate Ray Specific values
-            float rayIntensity = (float)1 - Mathf.Clamp(((float)bestRays[i].reflections * 1.1f / (float)rayStats.MaxReflections), 0f, 1f);
+            float rayIntensity = 1f - Mathf.Clamp(((float)bestRays[i].reflections * 1.1f / (float)rayStats.MaxReflections), 0f, 1f);
             float rayOcclusion = ((float)bestRays[i].occlusions / (float)rayStats.MaxOcclusions);
-            
             //Adding to emitter
             intensity += rayIntensity;
             occlusion += rayOcclusion;
